refactor(enumerators): convert ItemQuality to enum class with free functions
- Convert ItemQuality from custom class to enum class : uint8_t
- Implement item_quality_to_string and string_to_item_quality free functions
- Implement get_item_quality_modifier free function
- Update all usages in database writing, item models, subitems, and Lua bindings
- Replace .toString(), .toUInt(), and .getModifier() calls with free functions
- Update Lua bindings to use namespace constants

+++ src/database/sqliteWriteFunctions.cpp
@@ -163,7 +163,7 @@ bool SaveItem(Item * item,
     args.push_back(ToString(item->condition));
     args.push_back(ToString(item->maxCondition));
     args.push_back(ToString(item->composition->vnum));
-    args.push_back(ToString(item->quality.toUInt()));
+    args.push_back(ToString(static_cast<uint8_t>(item->quality)));
     args.push_back(ToString(item->flags));
     if (!transaction)
     {
